<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
-->
### What does this PR do?

<!--
* A brief description of the change being made with this pull request.
* If the description here cannot be expressed in a succinct form, consider
  opening multiple pull requests instead of a single one.
-->

Increase startup time threshold in Serverless Cold Start Integration test from 20s to 25s.

### Motivation

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

Several failures in the past week show that the test is flaky. There has been no noticeable increase in the median startup time in the past several weeks. Though it's likely the variance has increased, resulting in the flakiness.

<img width="887" alt="Screenshot 2024-07-25 at 10 22 55 AM" src="https://github.com/user-attachments/assets/53593bf3-326c-430c-9bf6-bd4a4869645e">

<img width="1802" alt="Screenshot 2024-07-25 at 10 23 56 AM" src="https://github.com/user-attachments/assets/14ad4202-b46d-41f1-96ed-568c0b86823e">

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

Threshold was decreased back to 20s on 7/16 following a temporary increase to 40s.

https://github.com/DataDog/datadog-agent/pull/27642

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->